Move Caff node test to the reorg test file, fix the finalized number source
